+++ php/Upload_Parts_Status.php
@@ -13,12 +13,19 @@ const EXPECTED_DELIVERY = 9;
 const RECEIVED_QTY		= 10;
 const INVOICE_DATE		= 11;
 const RETURNED_QTY		= 12;
+const SHOPLOCATION      = 13;
 
 function Upload_Parts_Status_CSV($parts_status_extract_file){
 
+        // Open the extract file and exit if not found.
+    if (($handle = fopen($parts_status_extract_file, "r")) === FALSE) {
+    	echo "Error in opening " . $parts_status_extract_file;
+    	exit;
+    }
+
     require('db_open.php');
 
-    	// Refresh the table of Parts Not Ordered
+    	//  Delete all records from the Parts Status table.
     $tsql = "DELETE FROM PartsStatusExtract";
 
     //echo $tsql;
@@ -30,16 +37,13 @@ function Upload_Parts_Status_CSV($parts_status_extract_file){
       exit;
     }
 
-    if (($handle = fopen($parts_status_extract_file, "r")) === FALSE) {
-    	echo "Error in opening " . $parts_status_extract_file;
-    	exit;
-    }
 
     $tsql = <<<strSQL
     		INSERT INTO PartsStatusExtract
-    			(RO_Num, Line, Part_Number, Part_Description, Part_Type,
-    			Vendor_Name, RO_Qty, Ordered_Qty, Order_Date,
-    			Expected_Delivery, Received_Qty, Invoice_Date, Returned_Qty)
+    			(RO_Num, Line, Part_Number, Part_Description,
+                Part_Type, Vendor_Name, RO_Qty, Ordered_Qty,
+                Order_Date, Expected_Delivery, Received_Qty,
+                Invoice_Date, Returned_Qty, Location)
     		VALUES
 strSQL;
 
@@ -83,11 +87,13 @@ strSQL;
 
     	$returned_quantity 	= $data[RETURNED_QTY];
 
+        $location           = "'" . Cleanup_Text($data[SHOPLOCATION]) . "'";
+
         $values = "(" . $ro_number . ", " . $line . ", " . $part_number . ", " .
                   $part_description . ", " . $part_type . ", " . $vendor_name . ", " .
     			  $ro_quantity . ", " . $ordered_quantity . ", " . $order_date . ", " .
     			  $expected_delivery . ", ". $received_quantity . ", " .
-    			  $invoice_date . "," . $returned_quantity . ")";
+    			  $invoice_date . "," . $returned_quantity ."," . $location . ")";
 
         $insert_sql = $tsql . $values;
     //	echo $insert_sql . '<br/><br/>';
